Make sure node user exists before using it
Sysdig Secure has been reporting a lot of the pods using this base image as "running as root", even though the user running the application in the pod is `node`.  Our Java base image does not have the same issue as the node image, and in that image we call the `adduser` command before using it.